test(test_general): move fixture to general module to test_general
